// src/errors/api-error-code.enum.ts
export enum ApiErrorCode {
  BAD_REQUEST = 'BAD_REQUEST',
  CONFLICT = 'CONFLICT',
  FORBIDDEN = 'FORBIDDEN',
  INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
  NOT_FOUND = 'NOT_FOUND',
  TOO_MANY_REQUESTS = 'TOO_MANY_REQUESTS',
  UNAUTHORIZED = 'UNAUTHORIZED',
  UNPROCESSABLE_ENTITY = 'UNPROCESSABLE_ENTITY',
  VALIDATION_FAILED = 'VALIDATION_FAILED',
}